POS_server/requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy==2.*
email_validator
python-decouple
gunicorn
python-dotenv
requests
authlib
six
pymysql
boto3
